Update Netezza driver with new method and Read re-implementation
Updated JustyBase.NetezzaDriver.Tests.csproj to change the
coverlet.collector package version from 6.0.3 to 6.0.4.

Added NewRowDescriptionStandardReceived method to NzConnection
class to check if the last response is of type RowDescriptionStandard.

Updated CheckIfRowsAreAvaiable method in NzDataReader class to
include a check for NewRowDescriptionStandardReceived and perform
the next step in the connection if the condition is met.

Added comments in NzDataReader class indicating the need for unit
tests and further implementation of columns.